feature: replace jinga steps by real steps
