+++ internal/workflows/steps/helpers_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/require"
+	"golang.hedera.com/solo-provisioner/internal/core"
 )
 
 func sudo(cmd *exec.Cmd) *exec.Cmd {
@@ -32,4 +33,11 @@ func TestRunCmd_Error(t *testing.T) {
 	_, err := runCmd("exit 42")
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "failed to execute bash command")
+}
+
+func cleanUpTempDir(t *testing.T) {
+	t.Helper()
+
+	err := os.RemoveAll(core.Paths().TempDir)
+	require.NoError(t, err)
 }
